+++ tools/legacy/sample_vpp/src/sample_vpp_utils.cpp
@@ -168,7 +168,8 @@ void PrintInfo(sInputParams* pParams, mfxVideoParam* pMfxParams, MFXVideoSession
                 Info.CropY,
                 Info.CropW,
                 Info.CropH);
-    msdk_printf(MSDK_STRING("Frame rate\t%.2f\n"), (mfxF64)Info.FrameRateExtN / Info.FrameRateExtD);
+    msdk_printf(MSDK_STRING("Frame rate\t%.2f\n"),
+                (double)((mfxF64)Info.FrameRateExtN / Info.FrameRateExtD));
     msdk_printf(MSDK_STRING("PicStruct\t%s\n"), PicStruct2Str(Info.PicStruct));
 
     Info = pMfxParams->vpp.Out;
@@ -179,7 +180,8 @@ void PrintInfo(sInputParams* pParams, mfxVideoParam* pMfxParams, MFXVideoSession
                 Info.CropY,
                 Info.CropW,
                 Info.CropH);
-    msdk_printf(MSDK_STRING("Frame rate\t%.2f\n"), (mfxF64)Info.FrameRateExtN / Info.FrameRateExtD);
+    msdk_printf(MSDK_STRING("Frame rate\t%.2f\n"),
+                (double)((mfxF64)Info.FrameRateExtN / Info.FrameRateExtD));
     msdk_printf(MSDK_STRING("PicStruct\t%s\n"), PicStruct2Str(Info.PicStruct));
 
     msdk_printf(MSDK_STRING("\n"));
@@ -249,7 +251,8 @@ void PrintInfo(sInputParams* pParams, mfxVideoParam* pMfxParams, MFXVideoSession
 
     msdk_printf(MSDK_STRING("IOpattern type               \t%s\n"),
                 IOpattern2Str(pParams->IOPattern));
-    msdk_printf(MSDK_STRING("Number of asynchronious tasks\t%hu\n"), pParams->asyncNum);
+    msdk_printf(MSDK_STRING("Number of asynchronious tasks\t%hu\n"),
+                (unsigned int)pParams->asyncNum);
     msdk_printf(MSDK_STRING("Time stamps checking         \t%s\n"),
                 pParams->ptsCheck ? MSDK_STRING("ON") : MSDK_STRING("OFF"));
